% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_calcCYERlimit.R
\name{calccyerobligation}
\alias{calccyerobligation}
\title{calccyerobligation}
\usage{
calccyerobligation(cmzpath, isbmmap)
}
\arguments{
\item{cmzpath}{full path to the cmz file , including file name
@export}
}
\description{
Calculate the CYER obligation based on the 2019 PST
}
\examples{

}
