% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_calmmoncyer.R
\name{calcmoncyer}
\alias{calcmoncyer}
\title{calmmoncyer}
\usage{
calcmoncyer(monthly.idi, cmzpath)
}
\arguments{
\item{monthly.idi}{data frame otput from the calcMonthly  function}

\item{cmzpath}{full path to the cmz file , including file name}
}
\description{
Read in  and clean up the CMZ file, a product of the mortality distribution tables program
}
\examples{

}
