% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_calcMonthly.R
\name{calcMonthly}
\alias{calcMonthly}
\title{calcMonthly}
\usage{
calcMonthly(
  stk = "ATN",
  fmap = NULL,
  pathdb = NULL,
  cas = NULL,
  pathcmz = NULL,
  tracing = FALSE
)
}
\arguments{
\item{stk}{= text string with stock name (e.g. "ATN")}

\item{fmap}{= data frame with fisheries mapping}

\item{pathdb}{= text string with path to the CAS database (need either this or the cas input below)}

\item{cas}{= data frame with CAS data, generated from databases (need either this or the pathdb input above)}

\item{pathcmz}{= text string with path to CMZ file}

\item{tracing}{= if TRUE, print various summaries}
}
\description{
Calculate monthly split of annual mortality rate
}
\examples{

}
