% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliter.R
\name{calmmoncyer}
\alias{calmmoncyer}
\title{calmmoncyer}
\usage{
calmmoncyer(casdf, cmz)
}
\arguments{
\item{casdf}{data frame otput from the readCASsplit function}

\item{cmz}{cmz file, output of  readcmzsplit}
}
\description{
Read in  and clean up the CMZ file, a product of the mortality distribution tables program
}
\examples{

}
