% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUB_utils.R
\name{include.dependencies}
\alias{include.dependencies}
\title{Internal handling of dependencies}
\usage{
include.dependencies()
}
\value{
Nothing returned.
}
\description{
Seems like this is needed in addition to the import list in the DESCRIPTION file, so keeping them all together in one place.
}
