% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_plotlydash.R
\name{plotcyerraster}
\alias{plotcyerraster}
\title{plotcyerraster}
\usage{
plotcyerraster(monthly.df.stock)
}
\arguments{
\item{monthly.df.stock}{data frame output from calcmoncyer, filtered by stock}
}
\description{
Plot output from calcMonthly function. See details there.
}
\examples{

}
