% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_plotMonthly.R
\name{plotMonthly}
\alias{plotMonthly}
\title{plotMonthly}
\usage{
plotMonthly(
  monthly.df,
  stk = "ATN",
  fgroup = "ISBM_CBC_Term_N",
  type = "basic"
)
}
\arguments{
\item{monthly.df}{data frame output from calcMonthly}

\item{stk}{stock to plot}

\item{fgroup}{fishery grouping to plot (depends on calcMonthly inputs!)}

\item{type}{plot type. For now, one of "basic" or "box"}
}
\description{
Plot output from calcMonthly function. See details there.
}
\examples{

}
