% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_readcasdbsplit.R
\name{readcasdbsplit}
\alias{readcasdbsplit}
\title{readcasdbsplit}
\usage{
readcasdbsplit(pathdb)
}
\arguments{
\item{pathdb}{path to the cas database from which CWT recoveries are to be extracted.}
}
\description{
Read in number of CWT by fishery an year based on query provided by Gayle
}
\examples{

}
