% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_writeRmd.R
\name{write_dash_raster}
\alias{write_dash_raster}
\title{write_dash_raster}
\usage{
write_dash_raster(fn, title, packs, code)
}
\arguments{
\item{monthly.df.stock.name}{name of data frame output from calcmoncyer, filtered by stock}

\item{obj_name}{stock acronym}
}
\value{
The Rmd header
}
\description{
write_dash_raster
}
