% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_writeRmd.R
\name{write_headr}
\alias{write_headr}
\title{Write headers for Rmd dashboard}
\usage{
write_headr(title)
}
\arguments{
\item{title}{Title of flexdashboard}
}
\value{
The Rmd header
}
\description{
Write headers for Rmd dashboard
}
