% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_writeRmd.R
\name{write_raster_chunks}
\alias{write_raster_chunks}
\title{Write R setup for Rmd dashboard}
\usage{
write_raster_chunks(monthly.df.name, obj_name)
}
\arguments{
\item{obj_name}{stock acronym}

\item{monthly.df.stock.name}{name of data frame output from calcmoncyer, filtered by stock}
}
\value{
The Rmd header
}
\description{
Write R setup for Rmd dashboard
}
