% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_writeRmd.R
\name{write_rmdcode}
\alias{write_rmdcode}
\title{Write R setup for Rmd dashboard}
\usage{
write_rmdcode(code)
}
\arguments{
\item{code}{Additional code one may want to include in the dashboard}
}
\value{
The Rmd header
}
\description{
Write R setup for Rmd dashboard
}
