% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_writeRmd.R
\name{write_setup}
\alias{write_setup}
\title{Write R setup for Rmd dashboard}
\usage{
write_setup(packs)
}
\arguments{
\item{packs}{vector containing the list of packages to be written in dashboard}
}
\value{
The Rmd header
}
\description{
Write R setup for Rmd dashboard
}
