% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_calcMonthly.R
\name{calcMonthly}
\alias{calcMonthly}
\title{calcMonthly}
\usage{
calcMonthly(
  fmap = NULL,
  pathdb = NULL,
  cas = NULL,
  tracing = FALSE,
  infill = TRUE
)
}
\arguments{
\item{fmap}{= data frame that matches 195 CAS fishery ids (CFILENO) in col 1 to a lookup grouping in col 2. Lookup groupings are used as output labels.}

\item{pathdb}{= text string with path to the CAS database (need either this or the cas input below)}

\item{cas}{= data frame with CAS data, generated from database (need either this or the pathdb input above)}

\item{tracing}{= if TRUE, print various summaries}

\item{infill}{= if TRUE, include months with 0 recoveries}
}
\description{
Calculate monthly distribution of tag recoveries for custom fishery groupings (loops through all stocks in the CAS input file).
}
\examples{

}
