% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliter.R
\name{readcmzsplit}
\alias{readcmzsplit}
\title{readcmzsplit}
\usage{
readcmzsplit(pathcmz)
}
\arguments{
\item{pathcmz}{path to the cmzfile .}
}
\description{
Read in  and clean up the CMZ file, a product of the mortality distribution tables program
}
\examples{

}
